+++ judgels-frontends/raphael/src/routes/uriel/contests/single/submissions/Bundle/ContestSubmissionRoutes.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import { withRouter } from 'react-router';
+
+const ContestSubmissionRoutes = () => <div>bundle submissions</div>;
+
+export default withRouter<any>(ContestSubmissionRoutes);
